requirements.txt
blinker==1.3
flask==0.10.1
flask-login==0.2.11
flask-oauthlib==0.8.0
flask-restful==0.3.1
flask-wtf==0.11
unidecode==0.4.17
gdata==2.0.18
google-api-python-client==1.2
